# routes/v1/openapi.yml
openapi: 3.0.0
info:
  description: Ourlabels API documentation
  version: 1.0.0
  title: Ourlabels API Documentation
  termsOfService: 'https://ourlabels.org/terms/'
  contact:
    email: ourlabels@gmail.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
paths:
  /add/annotations:
    post:
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAnnotationsRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /add/project:
    post:
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddProjectRequestBody'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /change/project:
    post:
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeProjectRequestBody'
      responses:
        '200': 
          description: OK Changed project
  /get/annotations:
    get:
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: offset
          schema:
            type: integer
            minimum: -1
            maximum: 1
          required: true
          example: 0
          description: Get most recent annotation for previous, next, or current image 
      responses:
        '200': 
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAnnotationsResponse'
        '400':
          description: Please select a project first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse400'
        '404':
          description: No un-annotated images found for project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse404'
...
  /get/projects:
    get:
      security:
        - BearerAuth: []
      responses:
        '200': 
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetProjectsResponse'
        '404':
          description: No projects found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse404'
...
  /join/project:
    post:
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: integer
      responses:
        '200':
          description: Successfully changed project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinProjectResponse'
        '400':
          description: Must include project_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse400'
        '404':
          description: No such project or unable to join
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse404'
  /leave/project:
    post:
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_id:
                  type: integer
      responses:
        '200':
          description: Successfully changed project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinProjectResponse'
        '400':
          description: Must include project_id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse400'
        '404':
          description: No such project or unable to join
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse404'
  /request/project:
    post:
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                project_name:
                  type: string
              required:
                - project_name
      responses:
        '200':
          description: Successfully added to requested users for project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: No such project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse400'
        '401':
          description: Unable to join project because refused from project or already requested
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse400'
...
  /update/project:
    post:
      security:
        - BearerAuth: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UpdateProjectRequest'
      responses:
        '200':
          description: OK
...
servers:
  - url: 'https://ourlabels.org/v1'
  - url: 'http://localhost:59003/v1'
components:
  schemas:
    AddAnnotationsRequestBody:
      type: object
      required: boxes
      properties:
        boxes:
          type: array
          items:
            $ref: '#/components/schemas/Box'
...
    AddProjectRequestBody:
      type: object
      required:
        - title
        - description
        - full_description
        - public
        - type
      properties:
        title:
          type: string
        description:
          type: string
        full_description:
          type: string
        public:
          type: boolean
        type:
          type: integer
          minimum: 0
...
    ChangeProjectRequestBody:
      type: object
      required: project_id
      properties:
        project_id: 
          type: integer
...
    Box:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        w:
          type: number
        h:
          type: number
        type:
          type: string
          example: mouse_body
      required:
        - x
        - y
        - w
        - h
        - type
    ProjectResponse:
      type: object
      properties:
        id: 
          type: integer
        title: 
          type: string
        description: 
          type: string
        full_description:
          type: string
        type:
          type: integer
        public:
          type: boolean
        featured:
          type: boolean
        owner:
          type: boolean
          description: If user is logged in, indicates if user is owner of project
        allowed:
          type: boolean
          description: If user is logged in, indicates if user is allowed to join project
        requested:
          type: boolean
          description: If user is logged in, indicates if user has requested access to project
        refused:
          type: boolean
          description: If user is logged in, indicates if user had request refused
        joined:
          type: boolean
          description: If user is logged in, indicates if user has joined project
        current_project:
          type: boolean
          description: If user is logged in, indicates if project is current project for user
...
    BoxResponse:
      type: object
      properties:
        x:
          type: number
        y:
          type: number
        width:
          type: number
        height:
          type: number
        type_key:
          type: string,
        justCreated:
          type: boolean,
        id:
          type: number      
    GetAnnotationsResponse:
      type: object
      properties:
        success: 
          type: boolean
        idx: 
          type: integer
        boxes:
          type: array
          items:
            $ref: '#/components/schemas/BoxResponse'
...
    GetProjectsResponse:
      type: object
      properties:
        success:
          type: boolean
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectResponse'
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
    ApiResponse400:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: array
          items: 
            type: object
    ApiResponse404:
      type: object
      properties:
        success:
          type: boolean
        error:
          type: string
...
    JoinProjectResponse:
      type: object
      properties:
        success:
          type: boolean
        joined:
          type: array
          items: 
            type: integer
        current_project:
          type: integer
...
    UpdateProjectRequest:
      type: object
      properties:
        files:
          type: array
          items:
            type: string
            format: binary
        project_id:
          description: The project id (integer row) being updated
          type: integer
        public:
          description: Whether the project will be visible to every user
          type: boolean
        owner:
          description: Either current owner or new owner, must be in allowed list
          type: string
        description:
          description: Short description of the project
          type: string
        full_description:
          description: Long description of the project
          type: string
        type:
          description: Integer id (row) type
          type: integer
        allowed:
          description: Array of usernames allowed to join the project
          type: array
          items: 
            type: string
        refused:
          description: Array of usernames to be disallowed
          type: array
          items: 
            type: string
        requested:
          description: New set of users who have not yet been refused or allowed
          type: array
          items: 
            type: string
        new:
          description: New sequences added to the project
          type: string
        update:
          description: Updated already existing sequence (maybe added new images or changed sequence name)
          type: string
        delete:
          description: Deleted sequences for project
          type: string
...